Bug 1123188 - [Flame][Dialer]It can't enter corresponding settings page when you tap "Voicemail" and "Fixed Dialing Numbers".
