+++ apps/settings/elements/call_iccs.html
@@ -22,7 +22,7 @@
       </ul>
     </div>
 
-    <script src="js/call_iccs.js"></script>
+    <panel data-path="panels/call_iccs/panel"></panel>
 
   </template>
 </element>
